updated methods in Board.h, fixed implementations of print and tetris

+++ the-latest-greatest-ai-game/the-latest-greatest-ai-game.cpp
@@ -8,7 +8,7 @@ int _tmain(int argc, _TCHAR* argv[])
 {
 	Board p = Board();
 	bool flag = true;
-	unsigned short col;
+	unsigned int col;
 	while (true){
 		std::cin >> col;
 		p.applyMove(col, flag);
